+++ revrt/routing/cli/point_to_feature.py
@@ -15,7 +15,7 @@ from revrt.routing.cli.base import (
     split_routes,
     RouteToDefinitionConverter,
 )
-from revrt.utilities.timing import log_runtime
+from revrt.utilities.monitoring import log_runtime
 from revrt.routing.utilities import map_to_costs
 from revrt.costs.config import parse_config
 from revrt.utilities.raster import integer_dimension_window
