+++ revrt/routing/cli/point_to_point.py
@@ -11,7 +11,7 @@ from revrt.routing.cli.base import (
     split_routes,
     RouteToDefinitionConverter,
 )
-from revrt.utilities.timing import log_runtime
+from revrt.utilities.monitoring import log_runtime
 from revrt.routing.utilities import map_to_costs
 from revrt.costs.config import parse_config
