+++ revrt/utilities/__init__.py
@@ -7,8 +7,6 @@ from .base import (
     expand_dim_if_needed,
     file_full_path,
     load_data_using_layer_file_profile,
-    log_mem,
-    log_array_backend,
     features_to_route_table,
     save_data_using_layer_file_profile,
     save_data_using_custom_props,
@@ -20,5 +18,11 @@ from .handlers import (
     num_feats_in_gpkg,
     gpkg_crs,
 )
-from .timing import log_runtime, elapsed_time_as_str
+from .monitoring import (
+    dask_performance_report,
+    log_mem,
+    log_array_backend,
+    log_runtime,
+    elapsed_time_as_str,
+)
 from .raster import rasterize, rasterize_shape_file, integer_dimension_window
